% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_link.R
\name{check_link}
\alias{check_link}
\title{Check link}
\usage{
check_link(x, website)
}
\arguments{
\item{x}{\code{x} An object, nominally a website loadded with \code{readr::read_lines} and passed to \code{XML::htmlParse()}.}

\item{website}{\code{website} Where the root of website is not given in the link, append the link portion to the root given by \code{website}.}
}
\value{
A character vector of size n>=1 which represent url links.
}
\description{
\code{check_link} check that a link is properly formatted .
}
\examples{

#library(dplyr)
#library(readr)
#library(XML)

#read_lines("http://www.google.co.uk") \%>\%
#htmlParse(encoding = "UTF8") \%>\%
#scrapeML::clean_string

}

