% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{Clean String}
\usage{
clean_string(x)
}
\arguments{
\item{x}{\code{x} An object, nominally a website loaded with \code{readr::read_lines} and passed to \code{XML::htmlParse()}.}
}
\value{
An object of class \code{character} with tags, etc removed, ready for building a document term matrix.
}
\description{
\code{clean_string} clean a string extracted using \code{XML::htmlParse()}
}
\examples{
library(dplyr)
library(readr)

read_lines("http://www.google.co.uk") \%>\%
htmlParse(encoding = "UTF8") \%>\%
scrapeML::clean_string
}

