% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dir.R
\name{parse_dir}
\alias{parse_dir}
\title{Parse dir}
\usage{
parse_dir(source_dir, dest_dir)
}
\arguments{
\item{source_dir}{\code{source_dir} Source directory in which downloaded html
files are stored.}

\item{dest_dir}{\code{dest_dir} Destination directory in which parsed
websites will be stored.}
}
\value{
Nothing is returned from the fucntion, but for each input html file a
  parsed version will be produced in the dest_dir folder
}
\description{
\code{clean_string} clean a string extracted using
  \code{XML::htmlParse()}
}
\examples{

#library(dplyr)
#library(readr)
#library(XML)

#read_lines("http://www.google.co.uk") \%>\% 
#htmlParse(encoding = "UTF8") \%>\%
#scrapeML::clean_string

}

