% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html.R
\name{parse_html}
\alias{parse_html}
\title{Clean String}
\usage{
parse_html(source_dir, dest_dir)
}
\arguments{
\item{x}{\code{x} An object, nominally a website loadded with \code{readr::read_lines} and passed to \code{XML::htmlParse()}.}
}
\value{
An object of class \code{character} with tags, etc removed, ready for building a document term matrix.
}
\description{
\code{clean_string} clean a string extracted using \code{XML::htmlParse()}
}
\examples{

library(dplyr)
library(readr)

read_lines("http://www.google.co.uk") \%>\% 
htmlParse(encoding = "UTF8") \%>\%
scrapeML::clean_string

}

