% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_match.R
\name{priority_match}
\alias{priority_match}
\title{Priority match}
\usage{
priority_match(x, patterns)
}
\arguments{
\item{x}{\code{x} A parsed html document. Nominally this will be the output
\code{XML::htmlParse(readr::read_Lines(x))}..}

\item{patterns}{\code{patterns} A character vector containing keywords in the
the order in which they should be searched for.}
}
\value{
If a match is found, the link string is returned. If no match is found
then \code{NA} is returned.
}
\description{
\code{priority_match} This is a scraping function used to
  sequentially check for the presence of keywords in an links in an html
  document. If the keyword of interest is found, then the function returns
  the link within which the keyword was found, otherwise it continues to
  search for the next keyword in the vector.
}
\examples{

#library(scrapeML)

#get_url("www.google.com",tempfile())

}

