% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_matrix.R
\name{tdm_matrix}
\alias{tdm_matrix}
\title{Get a Matrix of Terms}
\usage{
tdm_matrix(source_dir, stemming = FALSE)
}
\arguments{
\item{x}{\code{x} An object, nominally a website loaded with \code{readr::read_lines} and passed to \code{XML::htmlParse()}.}
}
\value{
An object of class \code{character} with tags, etc removed, ready for building a document term matrix.
}
\description{
\code{tdm_matrix}
}
\examples{

library(dplyr)
library(readr)

read_lines("http://www.google.co.uk") \%>\% 
htmlParse(encoding = "UTF8") \%>\%
scrapeML::clean_string

}

